Otro commit

+++ src/main/java/com/example/uf2_nf2_pr2_kilian_edgar/services/ComentarioService.java
@@ -19,6 +19,10 @@ public class ComentarioService {
     }
 
     public String deleteComentario(long id) {
+        Comentario comentario = comentarioRepository.findById(id);
+        if (comentario == null) {
+            return "Comentario doesn't exist";
+        }
         comentarioRepository.deleteById(id);
         return "Comentario deleted succesfully";
     }
